backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-redis==5.3.0
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
redis==4.6.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
